requirements.txt
pydub>=0.25.1
ffmpeg-python>=0.2.0
openai>=1.0.0
python-dotenv>=1.0.0
pydantic>=2.0.0
pytest>=7.0.0
black>=23.0.0
streamlit>=1.22.0
numpy>=1.22.0
tqdm>=4.65.0
loguru>=0.7.0
openai-whisper>=20231117
click>=8.1.3
langfuse>=2.0.0
